requirements.txt
requests==2.26.0
termcolor==1.1.0
python-dotenv==1.0.0